+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-### Version 7.0
-* Addition of Topic Metrics, porting to Extensions 2.+ framework
+### Version 7.0.0
+* Porting to appd-exts-commons 2.1.0
+* Support for Topic Metrics
 * config file extension changed from yaml to yml
-* config.yml changes to metric alias and formatting.
+* config.yml changes with respect to metrics.
 * ActiveChannelsCount metric added
+* monitor.xml changes
 
 ### Version 2.0
 * Extension revamp, multi threaded metric collection, configurable metrics.